+++ app/src/renderer/components/common/TmSessionHardware.vue
@@ -12,21 +12,15 @@
         <hardware-state
           v-if="status == 'connect'"
           icon="usb"
-          value="Please plug in your Ledger Wallet"
-          @click.native="setStatus('detect')"
+          value="Please plug in your Ledger Nano S and open the Cosmos app"
+          @click.native="connectLedger()"
         />
         <hardware-state
           v-if="status == 'detect'"
           :spin="true"
           icon="rotate_right"
-          value="Detecting your Ledger Wallet"
-          @click.native="setStatus('success')"
-        />
-        <hardware-state
-          v-if="status == 'success'"
-          icon="check_circle"
-          value="Ledger Wallet successfully loaded"
-          @click.native="onSubmit"
+          value="Connecting..."
+          @click.native="setStatus('connect')"
         />
       </div>
       <div class="tm-session-footer" />
@@ -50,13 +44,15 @@ export default {
     setStatus(value) {
       this.status = value
     },
-    onSubmit() {
-      this.$store.commit(`setModalSession`, false)
-      this.$store.commit(`notify`, {
-        title: `Welcome back!`,
-        body: `You are now signed in to your Cosmos account.`
-      })
-      this.$store.dispatch(`signIn`, { password: this.fields.signInPassword })
+    async connectLedger() {
+      this.setStatus(`detect`)
+      const connected = await this.$store.dispatch(`connectLedgerApp`)
+      if (connected) {
+        this.$store.commit(`notify`, {
+          title: `Connection succesful`,
+          body: `You are now signed in to your Cosmos account with your Ledger.`
+        })
+      }
     }
   }
 }