+++ app/src/renderer/connectors/lcdClient.js
@@ -24,8 +24,6 @@ const Client = (axios, remoteLcdURL) => {
     }
   }
 
-  const fetchAccount = argReq(`GET`, `/auth/accounts`)
-
   return {
     // meta
     lcdConnected: function() {
@@ -39,13 +37,13 @@ const Client = (axios, remoteLcdURL) => {
 
     // coins
     send: argReq(`POST`, `/bank/accounts`, `/transfers`),
-    queryAccount(address) {
+    queryAccount: function(address) {
       const emptyAccount = {
         coins: [],
         sequence: `0`,
         account_number: `0`
       }
-      return fetchAccount(address)
+      return req(`GET`, `/auth/accounts/${address}`)()
         .then(res => {
           return res.value || emptyAccount
         })